I can set null
The validation if the field IsNull or an empty string is invalid, because the dataset doens't work if I want to clear some field in my table, so, take all the dataset structure, except the invisible one.

The JSON to DataSet already brief null values and clear the fields when this ocurrs.